Replace Conduit w/optional flow+inlet+throttle
This gets rid of a lot of redundant code+tests, and opens up a fast
path for simple event handlers without any objects being created.
(Other than closures.)  You can now just call a source with a sink,
and that's all you need to do to have a subscription.  If the source
is async/unpausable and doesn't use any other streams, no extra flows
or throttles get created.

Note: this adds a lot of new APIs without any proper documentation,
and some of the old documentation is probably still out of date.