Overhaul share() operator backpressure management
It now only resumes the uplink if all downlinks are ready,
and pauses otherwise.  When a paused connection is closed,
the uplink will resume if that was the only paused connection.
(And there's some optimization for the common cases, where most
connections don't have inlets and thus don't have to be looped
over.)