Stream/signal unification part 2 - Sources!
Signals are now usable as Sources.  The old Source type is now
Producer, and the new Source type is a union of Signals and Producers.

This was necessary to avoid overloading shenanigans and incorrect
type inference for things that accepted Sources.  Going forward,
functions should usually *take* Sources and *return* Producers.
Producer is a narrower interface so TS is generally happy with
this approach.