Split until into until()/to()/next()
The until() function now always waits for a truthy value,
regardless of source type, while the new next() function
waits for a new or changed value.  Support for promises
was dropped, as we have to() and fromPromise() and even
start() -- plenty of ways to wait for a promise.

(Also, next() doesn't auto-convert functions to signals, so
it can be used with streams without pulling in the signals
framework as a requirement.)