Refactor demand monitoring for signals
A cell is "in demand" if it has subscribers or a queue set.
Changes in demand are detected when a queue is added or deleted,
or a cell gains or loses its sole subscriber, and is added to
an async batch to handle activating or deactivating resources
assoicated with that cell.

This refactoring is part of the groundwork for supporting the
use of both jobs and React-style `use` hooks inside of signals,
and to create a kind of persistent rule that doesn't need to
be destroyed and recreated when its enclosing parent detects
changes.