Improved demand monitoring and constant folding
Cells now subscribe to their dependencies *before* the
dependency catches up or recalculates, so that recalculation
always knows if the cell is observed/subscribed.  This means
constant folding is now an active process, rather than a
passive one (due to not subscribing until after recalc).
So now constant folding is recursive and eager, instead of
waiting for subscribers to read the constant again.  (This
also ensures a rapid drop in demand for the constant
signal(s), triggering cleanups sooner than would otherwise
happen.)