Simplify rule/job creation
Since cell jobs constantly restart, there's no point
binding them to the root, as they'll be unbound at the
first restart anyway.  We have to assume that rules
will be GC'd by their jobs, and then the lack of
demand will asynchronously GC the rest.

Also, our new ambient context means we don't need to
create a job for a rule before it's actually used: we
can just mark the cell stateful and leave it be.

Last and least, there's no reason to update a rule's
"changed" timestamp; that's a holdover from when we
thought it'd be possible to have inter-rule dependencies
or tiering via rules writing values.  In the present
system, a rule should never be depended upon by anything.